userspace: move udp echo server to proxier_test.go
